requirements.txt
python-dotenv
huggingface_hub
langchain
langchain_core
langchain_fireworks
langchain_huggingface
langgraph
fastapi
pyaudio
pypdf
streamlit
tokenizers

## OpenVoice requirements ##
librosa
faster-whisper
pydub
wavmark
numpy
eng_to_ipa
inflect
unidecode
whisper-timestamped
openai
pypinyin
cn2an
jieba
gradio
langid
